% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patients.R
\docType{data}
\name{patients}
\alias{patients}
\title{Patients data}
\format{
\subsection{\code{patients}}{

A data frame with 3 rows and 2 columns:
\describe{
\item{id}{ID}
\item{age}{Age}
}
}
}
\usage{
patients
}
\description{
Made up data for illustration purposes.
}
\keyword{datasets}
