% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{square}
\alias{square}
\title{Squares an input number}
\usage{
square(x)
}
\arguments{
\item{x}{a numeric value}
}
\value{
the squared value \code{x}
}
\description{
Squares an input number
}
\examples{
square(10)
}
