% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patients.R
\docType{data}
\name{patients}
\alias{patients}
\title{Patient data}
\format{
## `patients`
A data frame with 3 rows and 2 columns:
\describe{
  \item{id}{id f patient}
  \item{age}{years of age}
}
}
\usage{
patients
}
\description{
Some patient data
}
\keyword{datasets}
