% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{square}
\alias{square}
\title{Squares Numbers}
\usage{
square(x)
}
\arguments{
\item{x}{a numeric atomic, vector, or list}
}
\value{
the square of \code{x}
}
\description{
Squares Numbers
}
\examples{
square(10)
}
